simplified middleware, made matcher more invocation-efficient
